Refactor. BAP Authorization
Changes done according to PR request.
