+++ local.env
@@ -2,5 +2,5 @@ BUYER_APP_PORT=2010
 BUYER_APP_ID=sample_mobility_bap
 BUYER_APP_URL=http://localhost:2010
 GATEWAY_URL=http://localhost:1010
-privateKey=6os/UInKiuVpP5Vbu2m2PcVRX4SSIR2GbhrxQi1Wesum4lf4U7b24JgxjMDjQc1Srsc2/JgbuqkM7TKVS/SYdg==
-publicKey=puJX+FO29uCYMYzA40HNUq7HNvyYG7qpDO0ylUv0mHY=
+PRIVATE_KEY=6os/UInKiuVpP5Vbu2m2PcVRX4SSIR2GbhrxQi1Wesum4lf4U7b24JgxjMDjQc1Srsc2/JgbuqkM7TKVS/SYdg==
+PUBLIC_KEY=puJX+FO29uCYMYzA40HNUq7HNvyYG7qpDO0ylUv0mHY=
