+++ src/controllers/OnSearchController.js
@@ -1,14 +1,15 @@
 import LoggingService from '../services/LoggingService';
 import SearchService from '../services/SearchService';
-import genericResponse from './GenericResponse';
+import genericResponse from '../utilities/GenericResponse';
 import authVerifier from '../utilities/SignVerify/AuthHeaderVerifier';
+import LookUpService from '../services/LookUpService';
 
 const BecknGateway = 'BG';
 const onSearch = async (req, res) => {
   const logger = LoggingService.getLogger('OnSearchController');
   logger.debug(`on_search called with ${JSON.stringify(req.body)}`);
 
-  const publicKey = await genericResponse.getPublicKey(BecknGateway);
+  const publicKey = await LookUpService.getPublicKey(BecknGateway); 
   authVerifier.authorize(req, publicKey).then(() => {
     logger.debug('Request Authorized Successfully.');
     SearchService.storeSearchResult(req.body);
